HUGE UPDATES TO EVERYTHING!
Many things have changed. THe 15 C decode.py program was renamed to
"decoder.py" and modified to give the user more options, including the
option to know the shift value, the option to select the second
most-likely or more word and other various improvements.

Encoder can now randomly select a shift value between 0 and 26 and gives
the user the option to know the shift value if the randomize option is
selected.

A NEW FILE WAS CREATED!
We made a file that asks the user whether or not they wish to encode or
decode so that the user does not need to enter either or those files to
do one of those actions.

These have been tested plenty, however I encourage you to test it
yourself as  there are now so many different combinations of conditions
that I may have missed something.

This is about as many of the changes as I can remember. Will have to
keep notes next time.